requirements.txt
newspaper4k
sqlalchemy
feedparser
lxml_html_clean
aiohttp